fix: Rename action-lookup to action-operations and refactor related tests
